--- web-react/src/pages/portfolio/games/games.css
@@ -1,43 +1,0 @@
-p {
-    font-size: medium;
-    letter-spacing: 0.5px;
-}
-
-h3 {
-    margin-bottom: 4px;
-}
-
-.games__row:hover {
-    color: black !important;
-}
-
-.games__row:hover h3 {
-    color: #5289B5;
-}
-
-.games__container {
-    margin: 48px 0;
-}
-
-.games__row {
-    display: flex;
-    align-items: flex-start;
-    pointer: cursor;
-}
-
-.games__contentImage__container {
-    width: 200px;
-    text-align: center;
-}
-
-.games__contentImage__image {
-    transform: scale(1.25);
-}
-
-.games__contentText__container {
-    margin-left: 32px;
-}
-
-.games__contentText__category {
-    color: grey;
-}
