--- web-react/src/pages/portfolio/productivity/productivity.css
@@ -1,42 +1,0 @@
-p {
-    font-size: medium;
-}
-
-h3 {
-    margin-bottom: 4px;
-}
-
-.productivity__row:hover {
-    color: black !important;
-}
-
-.productivity__row:hover h3 {
-    color: #5289B5;
-}
-
-.productivity__container {
-    margin: 48px 0;
-}
-
-.productivity__row {
-    display: flex;
-    align-items: flex-start;
-    pointer: cursor;
-}
-
-.productivity__contentImage__container {
-    width: 200px;
-    text-align: center;
-}
-
-.productivity__contentImage__image {
-    transform: scale(1.25);
-}
-
-.productivity__contentText__container {
-    margin-left: 32px;
-}
-
-.productivity__contentText__category {
-    color: grey;
-}
